requirements.txt
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.1
joblib==1.3.2
kiwisolver==1.4.4
markdown-it-py==3.0.0
matplotlib==3.7.2
mdurl==0.1.2
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.25.2
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pycodestyle==2.11.0
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
regex==2023.8.8
rich==13.5.2
six==1.16.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.7.1
wordcloud==1.9.2
